+++ src/main/java/school/hei/patrimoine/possession/Possession.java
@@ -13,5 +13,5 @@ public sealed abstract class Possession permits
   protected final Instant t;
   protected final int valeurComptable;
 
-  public abstract float valeurComptableFuture(Instant tFutur);
+  public abstract double valeurComptableFuture(Instant tFutur);
 }
